+++ src/main/java/teamroots/embers/mixin/jaopca/EmbersCompatModuleMixin.java
@@ -1,6 +1,5 @@
 package teamroots.embers.mixin.jaopca;
 
-import thelm.jaopca.compat.embers.EmbersCompatModule;
 import net.minecraft.item.Item;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -8,8 +7,7 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 
 import teamroots.embers.register.ItemRegister;
 
-
-@Mixin(value = EmbersCompatModule.class, remap = false)
+@Mixin(targets = "thelm.jaopca.compat.embers.EmbersCompatModule", remap = false)
 public class EmbersCompatModuleMixin {
     @Redirect(method = "onInit", at = @At(value = "FIELD", target = "Lteamroots/embers/RegistryManager;stamp_bar:Lnet/minecraft/item/Item;"))
     private Item stamp_bar() {