+++ src/main/java/teamroots/embers/mixin/jaopca/EmbersModuleMixin.java
@@ -4,9 +4,8 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import teamroots.embers.config.ConfigMisc;
-import thelm.jaopca.compat.embers.EmbersModule;
 
-@Mixin(value = EmbersModule.class, remap = false)
+@Mixin(targets = "thelm.jaopca.compat.embers.EmbersModule", remap = false)
 public class EmbersModuleMixin {
     @Redirect(method = "onInit", at = @At(value = "FIELD", target = "Lteamroots/embers/ConfigManager;melterOreAmount:I"))
     private int onInit() {