+++ src/main/java/teamroots/embers/tileentity/TileEntityMixerBottom.java
@@ -184,7 +184,11 @@ public class TileEntityMixerBottom extends TileEntity implements ITileEntityBase
         TileEntityMixerTop top = (TileEntityMixerTop) world.getTileEntity(pos.up());
         isWorking = false;
         if (top != null) {
-            upgrades = UpgradeUtil.getUpgrades(world, pos, EnumFacing.VALUES);
+            upgrades = UpgradeUtil.getUpgrades(world, pos.up(), EnumFacing.VALUES);
+            // if we can't find any upgrades, check the bottom
+            if (upgrades.isEmpty()) {
+                upgrades = UpgradeUtil.getUpgrades(world, pos, EnumFacing.VALUES);
+            }
             UpgradeUtil.verifyUpgrades(this, upgrades);
             if (UpgradeUtil.doTick(this, upgrades))
                 return;